Update toolchain to gcc 7.3
Updating a few other dependencies as they don't compile with gcc 7.3 otherwise.

+++ source/thrift/build.sh
@@ -33,9 +33,10 @@ download_dependency $LPACKAGE "${LPACKAGE_VERSION}.tar.gz" $THIS_DIR
 if needs_build_package ; then
   header $PACKAGE $PACKAGE_VERSION
 
-  BOOST_ROOT=$BUILD_DIR/boost-$BOOST_VERSION
-  ZLIB_ROOT=$BUILD_DIR/zlib-$ZLIB_VERSION
-  LIBEVENT_ROOT=$BUILD_DIR/libevent-$LIBEVENT_VERSION
+  BISON_ROOT="${BUILD_DIR}"/bison-"${BISON_VERSION}"
+  BOOST_ROOT="${BUILD_DIR}"/boost-"${BOOST_VERSION}"
+  OPENSSL_ROOT="${BUILD_DIR}"/openssl-"${OPENSSL_VERSION}"
+  ZLIB_ROOT="${BUILD_DIR}"/zlib-"${ZLIB_VERSION}"
 
   # If we build in local dev mode, use the bundled OpenSSL
   if [[ "$PRODUCTION" -eq "0" || "$OSTYPE" == "darwin"* ]]; then
@@ -54,30 +55,54 @@ if needs_build_package ; then
     wrap glibtoolize --copy
     wrap autoconf
   fi
-  JAVA_PREFIX=${LOCAL_INSTALL}/java PY_PREFIX=${LOCAL_INSTALL}/python \
-    wrap ./configure --with-pic --prefix=${LOCAL_INSTALL} \
+
+  PATH="${BISON_ROOT}"/bin:"${PATH}" \
+    PY_PREFIX="${LOCAL_INSTALL}"/python \
+    wrap ./configure \
+    --with-pic \
+    --prefix="${LOCAL_INSTALL}" \
+    --enable-tutorial=no \
     --with-c_glib=no \
-    --with-php=no --with-java=no --with-perl=no --with-erlang=no --with-csharp=no \
-    --with-ruby=no --with-haskell=no --with-erlang=no --with-d=no \
-    --with-boost=${BOOST_ROOT} \
-    --with-zlib=${ZLIB_ROOT} \
-    --with-libevent=${LIBEVENT_ROOT} \
+    --with-php=no \
+    --with-java=no \
+    --with-perl=no \
+    --with-erlang=no \
+    --with-csharp=no \
+    --with-ruby=no \
+    --with-haskell=no \
+    --with-erlang=no \
+    --with-d=no \
+    --with-boost="${BOOST_ROOT}" \
+    --with-zlib="${ZLIB_ROOT}" \
     --with-nodejs=no \
     --with-lua=no \
-    --with-go=no --with-qt4=no --with-libevent=no ${PIC_LIB_OPTIONS:-} $OPENSSL_ARGS
-  wrap make   # Build fails with -j${BUILD_THREADS}
-  wrap make install
+    --with-go=no \
+    --with-qt4=no \
+    --with-libevent=no \
+    ${PIC_LIB_OPTIONS:-} \
+    ${OPENSSL_ARGS} \
+    ${CONFIGURE_FLAG_BUILD_SYS}
+  # The error code is zero if one or more libraries can be built. To ensure that C++
+  # and python libraries are built the output should be checked.
+  if ! grep -q "Building C++ Library \.* : yes" "${BUILD_LOG}"; then
+    echo "Thrift cpp lib configuration failed."
+    exit 1
+  fi
+  if ! grep -q "Building Python Library \.* : yes" "${BUILD_LOG}"; then
+    echo "Thrift python lib configuration failed."
+    exit 1
+  fi
+  wrap make install # Thrift 0.9.0 doesn't build with -j${BUILD_THREADS}
   cd contrib/fb303
   rm -f config.cache
   chmod 755 ./bootstrap.sh
-  wrap ./bootstrap.sh --with-boost=${BOOST_ROOT}
+  wrap ./bootstrap.sh --with-boost="${BOOST_ROOT}"
   wrap chmod 755 configure
-  CPPFLAGS="-I${LOCAL_INSTALL}/include" PY_PREFIX=${LOCAL_INSTALL}/python wrap ./configure \
-    --with-boost=${BOOST_ROOT} \
-    --with-java=no --with-php=no --prefix=${LOCAL_INSTALL} \
-    --with-thriftpath=${LOCAL_INSTALL} $OPENSSL_ARGS
-  wrap make -j${BUILD_THREADS}
-  wrap make install
+  CPPFLAGS="-I${LOCAL_INSTALL}/include" PY_PREFIX="${LOCAL_INSTALL}"/python wrap ./configure \
+    --with-boost="${BOOST_ROOT}" \
+    --with-java=no --with-php=no --prefix="${LOCAL_INSTALL}" \
+    --with-thriftpath="${LOCAL_INSTALL}" ${OPENSSL_ARGS}
+  wrap make -j"${BUILD_THREADS}" install
 
   footer $PACKAGE $PACKAGE_VERSION
 fi
